Add data import/export functionality for CVS, XLS, and HDF5
Two buttons for import/export were added to each of the three tables.
Three file formats can be used for importing and exporting data: CSV,
XLS, and HDF5.

There is no proper error handling for the file readers and writers, so
it is advised to import only files that were previously exported by this
control.

In addition to adding the new features, a lot of code was refactored.